commit changes of past 2 days, see changelog for details most important changes:  - added MSXPrinterPort  - PSG sounds louder
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@623 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/Mixer.hh
@@ -53,7 +53,7 @@ class Mixer
 		int nbAllDevices;
 		int nbDevices[NB_MODES];
 		std::vector<SoundDevice*> devices[NB_MODES];
-		std::vector<short*> buffers[NB_MODES];
+		std::vector<int*> buffers[NB_MODES];
 		
 		short* mixBuffer;
 		int samplesLeft;
